+++ app.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const Homey = require('homey');
-const server = require('./node-sonos-http-api-master/server');
 const sonos_api = require('./node-sonos-http-api-master/sonos_api');
 
 class SonosSay extends Homey.App {
