+++ app.json
@@ -1,6 +1,6 @@
 {
   "id": "com.svipler.athom.sonos-say",
-  "version": "3.1.5",
+  "version": "4.0.0",
   "compatibility": ">=5.0.0",
   "sdk": 3,
   "name": {
