+++ node-sonos-http-api-master/lib/sonos-http-api.js
@@ -94,67 +94,6 @@ function HttpAPI(discovery, settings) {
       });
   };
 
-  this.requestHandler = function (req, res) {
-    if (req.url === '/favicon.ico') {
-      res.end();
-      return;
-    }
-
-    if (discovery.zones.length === 0) {
-      const msg = 'No system has yet been discovered. Please see https://github.com/jishi/node-sonos-http-api/issues/77 if it doesn\'t resolve itself in a few seconds.';
-      logger.error(msg);
-      sendResponse(500, { status: 'error', error: msg });
-      return;
-    }
-
-    const params = req.url.substring(1).split('/');
-
-    // parse decode player name considering decode errors
-    let player;
-    try {
-      player = discovery.getPlayer(decodeURIComponent(params[0]));
-    } catch (error) {
-      logger.error(`Unable to parse supplied URI component (${params[0]})`, error);
-      return sendResponse(500, { status: 'error', error: error.message, stack: error.stack });
-    }
-
-    const opt = {};
-
-    if (player) {
-      opt.action = (params[1] || '').toLowerCase();
-      opt.values = params.splice(2);
-    } else {
-      player = discovery.getAnyPlayer();
-      opt.action = (params[0] || '').toLowerCase();
-      opt.values = params.splice(1);
-    }
-
-    function sendResponse(code, body) {
-      var jsonResponse = JSON.stringify(body);
-      res.statusCode = code;
-      res.setHeader('Content-Length', Buffer.byteLength(jsonResponse));
-      res.setHeader('Content-Type', 'application/json;charset=utf-8');
-      res.write(Buffer.from(jsonResponse));
-      res.end();
-    }
-
-    opt.player = player;
-    Promise.resolve(handleAction(opt))
-      .then((response) => {
-        if (!response || response.constructor.name === 'IncomingMessage') {
-          response = { status: 'success' };
-        } else if (Array.isArray(response) && response.length > 0 && response[0].constructor.name === 'IncomingMessage') {
-          response = { status: 'success' };
-        }
-
-        sendResponse(200, response);
-      }).catch((error) => {
-        logger.error(error);
-        sendResponse(500, { status: 'error', error: error.message, stack: error.stack });
-      });
-  };
-
-
   function handleAction(options) {
     var player = options.player;
