+++ node-sonos-http-api-master/package.json
@@ -13,16 +13,13 @@
   },
   "dependencies": {
     "anesidora": "^1.2.0",
-    "basic-auth": "~1.1.0",
-    "fuse.js": "^6.4.1",
-    "html-entities": "^1.2.1",
+    "fuse.js": "^6.4.6",
+    "html-entities": "^1.4.0",
     "json5": "^0.5.1",
-    "mime": "^1.4.1",
     "music-metadata": "^1.1.0",
-    "node-static": "^0.7.9",
+    "node-static": "^0.7.11",
     "request-promise": "~1.0.2",
-    "sonos-discovery": "https://github.com/jishi/node-sonos-discovery/archive/v1.7.3.tar.gz",
-    "wav-file-info": "0.0.8"
+    "sonos-discovery": "https://github.com/jishi/node-sonos-discovery/archive/v1.7.3.tar.gz"
   },
   "engines": {
     "node": ">=4.0.0",
@@ -33,6 +30,6 @@
   "devDependencies": {
     "eslint": "^4.8.0",
     "eslint-config-airbnb-base": "^12.0.1",
-    "eslint-plugin-import": "^2.7.0"
+    "eslint-plugin-import": "^2.22.1"
   }
 }
